feat: remove deprecated unused segment status crontab
