+++ src/console/src/types.rs
@@ -158,7 +158,6 @@ pub mod interface {
     use crate::types::state::ProposalId;
     use candid::CandidType;
     use junobuild_shared::types::core::Hash;
-    use junobuild_shared::types::cronjob::CronJobs;
     use junobuild_storage::types::config::StorageConfig;
     use serde::{Deserialize, Serialize};
 
@@ -167,11 +166,6 @@ pub mod interface {
         Satellite,
         MissionControl,
         Orbiter,
-    }
-
-    #[derive(CandidType, Deserialize)]
-    pub struct CronJobsArgs {
-        pub cron_jobs: CronJobs,
     }
 
     #[derive(CandidType, Deserialize)]
