+++ src/mission_control/src/lib.rs
@@ -3,7 +3,6 @@ mod controllers;
 mod guards;
 mod impls;
 mod memory;
-mod mgmt;
 mod monitoring;
 mod segments;
 mod store;
@@ -19,11 +18,8 @@ use crate::controllers::satellite::{
     remove_satellite_controllers as remove_satellite_controllers_impl, set_satellite_controllers,
 };
 use crate::controllers::store::get_controllers;
-use crate::guards::{
-    caller_is_user_or_admin_controller, caller_is_user_or_admin_controller_or_juno,
-};
+use crate::guards::caller_is_user_or_admin_controller;
 use crate::memory::{get_memory_upgrades, init_runtime_state, init_stable_state, STATE};
-use crate::mgmt::status::collect_statuses;
 use crate::segments::orbiter::{
     attach_orbiter, create_orbiter as create_orbiter_console,
     create_orbiter_with_config as create_orbiter_with_config_console, delete_orbiter,
@@ -52,7 +48,7 @@ use crate::types::state::{
 use candid::Principal;
 use ciborium::into_writer;
 use ic_cdk::api::call::{arg_data, ArgDecoderConfig};
-use ic_cdk::{id, storage, trap};
+use ic_cdk::{storage, trap};
 use ic_cdk_macros::{export_candid, init, post_upgrade, pre_upgrade, query, update};
 use ic_ledger_types::{Tokens, TransferArgs, TransferResult};
 use icrc_ledger_types::icrc1::transfer::TransferArg;
@@ -61,11 +57,9 @@ use junobuild_shared::ledger::icrc::icrc_transfer_token;
 use junobuild_shared::ledger::types::icrc::IcrcTransferResult;
 use junobuild_shared::mgmt::cmc::top_up_canister;
 use junobuild_shared::mgmt::ic::deposit_cycles as deposit_cycles_shared;
-use junobuild_shared::types::interface::{
-    DepositCyclesArgs, MissionControlArgs, SetController, StatusesArgs,
-};
+use junobuild_shared::types::interface::{DepositCyclesArgs, MissionControlArgs, SetController};
 use junobuild_shared::types::state::{
-    ControllerId, ControllerScope, Controllers, OrbiterId, SatelliteId, SegmentsStatuses,
+    ControllerId, ControllerScope, Controllers, OrbiterId, SatelliteId,
 };
 use junobuild_shared::types::state::{Metadata, UserId};
 use junobuild_shared::upgrade::write_pre_upgrade;
@@ -402,15 +396,6 @@ fn version() -> String {
 // ---------------------------------------------------------
 // Observatory
 // ---------------------------------------------------------
-
-#[deprecated(
-    since = "0.0.14",
-    note = "Deprecated with the introduction of monitoring features that include auto top-up capabilities."
-)]
-#[update(guard = "caller_is_user_or_admin_controller_or_juno")]
-async fn status(config: StatusesArgs) -> SegmentsStatuses {
-    collect_statuses(&id(), &config).await
-}
 
 #[deprecated(
     since = "0.0.14",
