--- src/mission_control/src/mgmt/constants.rs
@@ -1,1 +1,0 @@
-pub const CYCLES_MIN_THRESHOLD: u64 = 500_000_000_000;
