--- src/mission_control/src/mgmt/mod.rs
@@ -1,2 +1,0 @@
-pub mod constants;
-pub mod status;
